Make sure two Org faces use the italic style rather than the main foreground
This was always the intent, as those faces exist in the
'doric-themes-italic-faces' but I left them behind in the
'doric-themes-main-foreground-only-faces' from when I originally set
up the themes.

Thanks to Rudolf Adamkovič for noting that 'org-latex-and-related' did
not stand out, as it should, and for then finding that
'org-inline-src-block' had the same problem. This was done in issue
16: <https://github.com/protesilaos/doric-themes/issues/16>.