Remove override for shr-text face
We do not need to do it this way. I had included faces there as a
brute force way to remove their colours. Eventually the list of
doric-themes-main-foreground-only-faces will be shorter.

This is in response to Marcus Kammer for telling me that nov-mode
buffers were not proportionately spaced by default. This was done in
issue 23: <https://github.com/protesilaos/doric-themes/issues/23>.